fix warning compose

+++ app/src/main/java/com/example/foodapps/presention/ui/recipe_list/RecipeListViewModel.kt
@@ -74,6 +74,6 @@ class RecipeListViewModel
     }
 
     fun onChangeCategoryScrollPosition(position: Int){
-        categoryScrollPosition = position
+        categoryScrollPosition = position.toFloat()
     }
 }
